simulation of a propagating sinusoidal striking a lossy dielectric medium

+++ fd1d/clang/test_1_5.c
@@ -18,8 +18,8 @@ typedef struct {
 
 
 float sinusoidal(int t, float ds, float freq) {
-    float dt = ds/6e8;  /* time step (s) */
-    return sin(2*M_PI*freq*dt*t);
+    float dt = ds/6e8f;  /* time step (s) */
+    return sinf(2*M_PI*freq*dt*t);
 }
 
 
@@ -28,7 +28,7 @@ void exfield(int t, int nx, float *ca, float *cb, float *ex, float *hy) {
     for (int i = 1; i < nx; i++)
         ex[i] = ca[i] * ex[i] + cb[i] * (hy[i-1] - hy[i]);
     /* put a sinusoidal wave at the low end */
-    ex[1] += sinusoidal(t, 0.01f, 700e6);
+    ex[1] += sinusoidal(t, 0.01f, 700e6f);
 }
 
 
@@ -38,7 +38,7 @@ void hyfield(int nx, float *ex, float *hy, float *bc) {
     ex[nx-1] = bc[3], bc[3] = bc[2], bc[2] = ex[nx-2];
     /* calculate the Hy field */
     for (int i = 0; i < nx-1; i++)
-        hy[i] += 0.5 * (ex[i] - ex[i+1]);
+        hy[i] += 0.5f * (ex[i] - ex[i+1]);
 }
 
 
@@ -48,10 +48,10 @@ tuple dielectric(int nx, float dt, float epsr, float sigma) {
     n.cb = (float*) calloc(nx, sizeof(*n.cb));
     for (int i = 0; i < nx; n.ca[i] = 1.0f, i++);
     for (int i = 0; i < nx; n.cb[i] = 0.5f, i++);
-    float eps0 = 8.854e-12;  /* vacuum permittivity (F/m) */
+    float eps0 = 8.854e-12f;  /* vacuum permittivity (F/m) */
     float epsf = dt*sigma/(2*eps0*epsr);
     for (int i = nx/2; i < nx; n.ca[i] = (1 - epsf)/(1 + epsf), i++);
-    for (int i = nx/2; i < nx; n.cb[i] = 0.5/(epsr*(1 + epsf)), i++);
+    for (int i = nx/2; i < nx; n.cb[i] = 0.5f/(epsr*(1 + epsf)), i++);
     return n;
 }
 
@@ -66,10 +66,10 @@ int main() {
 
     float bc[4] = {0.0f};
 
-    float ds = 0.01;  /* spatial step (m) */
-    float dt = ds/6e8;  /* time step (s) */
-    float epsr = 4.0;  /* relative permittivity */
-    float sigma = 0.04;  /* conductivity (S/m) */
+    float ds = 0.01f;  /* spatial step (m) */
+    float dt = ds/6e8f;  /* time step (s) */
+    float epsr = 4.0f;  /* relative permittivity */
+    float sigma = 0.04f;  /* conductivity (S/m) */
     tuple n = dielectric(nx, dt, epsr, sigma);
 
     clock_t stime = clock();
